infoframe done fix #24
