+++ src/routes/[slug]/+page.svelte
@@ -1,12 +1,8 @@
 <script>
+    import ArtikelInfoframe from '../../components/artikel-infoframe.svelte';
+
     /** @type {import('./$types').PageData} */
     export let data;
-
-    const dateFormat = {
-        day: 'numeric',
-        month: 'long',
-        year: 'numeric'
-    };
     
     // Check if the data has been received and is an array
     console.log("Received data in +page.svelte:", data);
@@ -14,11 +10,8 @@
 </script>
 
 {#if post}
+    <ArtikelInfoframe post={post}></ArtikelInfoframe>
     <!-- @html means: there is html in this string, render it -->
-    <p>{(new Date(post.date)).toLocaleDateString("nl-NL", dateFormat)}</p>
-    <h3>{@html post.title.rendered}</h3> 
-    <p>{@html post.excerpt.rendered}</p>
-    <img src={post.yoast_head_json.og_image[0].url} alt="Artikel afbeelding">
     <p>{@html post.content.rendered}</p>
 
     <div>
